Global constant initialisation constraint
Summary:
In a previous diff (D39128345) we introduced support for global constants to Hips, using one-directed constraint propagation.

For example, we were able to derive for the following program (which type-checks, but fails at runtime)

  <?hh

  const dict<string, mixed> DICT = dict['a' => 42];

  function main(): void {
    DICT['b'];
  }

the output

  Summary after closing and simplifying for \DICT:
  [3:27-31] [Constant]:
    shape(?'a' => int, ?'b' => mixed)
  [3:34-49] [Allocation]:
    shape('a' => int)
  Summary after closing and simplifying for \main:

In the future, we would like to deduce the more refined output (`a` non-optional, since constants can`t be modified (!))

  [3:27-31] [Constant]:
    shape('a' => int, ?'b' => mixed)

This more precise analysis requires an additional constraint propagation in the reverse direction (from `\DICT` to `main`).

In this diff, we pave the road, and introduce a new inter-procedural constraint `ConstantInitial of 'a`, which we will use for further analysis.

Reviewed By: madgen

Differential Revision: D39167629

fbshipit-source-id: b501d57b8e9234ab21cfd55da5f0c930acd3aeb4