Modifiying module type Inter and instantiating shape-like-dict
Summary:
This diff is twofold:

1:
We modify the module type `Inter` in `hips_types.ml` in the following way:

- We replace `val substitute_inter_intra` with the two functions `val substitute_inter_intra_forwards` and `val substitute_inter_intra_backwards`.
- We modify the type of  `val is_same_entity : entity -> intra_entity -> bool` to `intra_entity -> intra_entity -> bool`.
- We require a new function ` val embed_entity : param_entity -> intra_entity`.

With this new data at hand, we extend the `substitute` function in `hips_solver` with a direction that is analogous to the one we used for global constants.

Note: we now do not copy all constraints anymore, but only the ones that use the relevant function argument.

2:
We modify `shape_analysis_solver.ml` such that it generates the right new data to instantiate a module of type `Inter` in `shape_analysis_hips.ml`.

Note: This diff in particular stops the reporting of empty shapes for parameters, a problem we previously encountered.

Reviewed By: madgen

Differential Revision: D38954803

fbshipit-source-id: bd30c984edd4394e500f737c5fca7626813175e2