Removing duplicated constraints / correcting fixpoint calculation
Summary:
To my surprise changing `max_iteration` in `shape_analysis_hips.ml` from `2` to e.g. `5` resulted in non-termination of the solve-analysis of `solve_constraints/recursion.php`.

In other words, there hadn't yet been a fixpoint reached after `2` iterations, and further iterations were not feasible, as they lead to an explosion in the number of constraints.

This diff addresses the issue by removing duplicated constraints in the closing procedure. In consequence termination with convergence to a fixpoint is achieved (we set divergence termination to `15` iterations) for  `solve_constraints/recursion.php`. Solving now seems remarkably quick.

Reviewed By: madgen

Differential Revision: D39328272

fbshipit-source-id: eb242a69eccb88afa4c88d3eed56925dabe9e46f