Entities for constant and identifier, and marker for constant
Summary:
We would like to extend the inter-procedural constraint solver for Hack (HIPS) to global constants.

Our long-term goal is to implement code-mods as specified in https://fb.quip.com/RUPbAVitrdoP.

This diff makes the following contributions:
- We introduce two new entities `const_entity`, `identifier_entity`
- We introduce a new marker `Constant`

In the future diff D39131395 we use both entities within newly defined inter-procedural constraints. In D39128345 we extend hips such that it globally uses this information to add subset constraint relations between constant and identifier entities. This is necessary to deduce the right shape type of a constant entity, which is reported by the analysis via the `Constant` marker.

Reviewed By: madgen

Differential Revision: D38869927

fbshipit-source-id: f47c6b68878377b0c2e90df70e39ae08b097e60b