Always return an aast::Program
Summary:
The lowerer is intended to always return an `aast::Program`, along with any parsing errors. This is important for e.g. IDE services, where we need to parse `$x->` accurately to run type inference and offer accurate code completion.

The return type of `lower()` is `Result<Program, String>` which breaks this convention. What it does today:

(1) If there's a parse error during the FFP construction, we return Ok(whole_program)

(2) If there's a 'missing syntax' error during lowering, we return Ok(valid_toplevel_defs)

(3) If there's a 'failwith' parsing error, we return Err("parsing error").

Both (1) and (2) are surprising as it's easy to assume a syntax error from the user would produce Err(). (3) is particularly bad because we discard the entire AAST.

Instead, change the type returned from the lower to always return a `aast::Program`, and write parse errors to `env` in all cases.

Reviewed By: aorenste

Differential Revision: D39227979

fbshipit-source-id: fe32312f2f88ca25a3620bbd686abbe8fe0bc2e9