Defer surprise check until after DV initializers
Summary:
Currently, event hooks can observe whether a parameter with a default value was
passed by the caller or not. If a unit test framework depends on this behavior,
it takes away HHVM's freedom for moving this initialization to other places,
such as a caller or a prologue. It is also desirable to be able to jump to the
function's main func entry translation after setting default arguments, instead
of skipping over func entry and jumping to the body, due to the translation
fragmentation issues. Since the main func entry needs to perform the surprise
check, it needs to happen after DV initializers.

This diff moves them from DV func entries to the Enter opcode.

Reviewed By: ricklavoie

Differential Revision: D38964459

fbshipit-source-id: f4486a5a5a26ae7620e395959a214c5e4348387e