Add glog::glog as a dependency of Folly (#9178)
Summary:
Pull Request resolved: https://github.com/facebook/hhvm/pull/9178

`glog` is a dependency of Folly but we did not explicit specify it for CMake, as a result, the order of libraries in the linker command is incorrect.

This diff set `glog::glog` as a dependency of Folly, so that CMake could topologically sort the libraries and create a linker command, in which the order of libraries is correct.

Reviewed By: Wilfred

Differential Revision: D39365630

fbshipit-source-id: 49429d71b3441dbcb7056cceab8793e3a18bc04f