Stop the thrift python converter from throwing with bad enums
Summary:
In thrift py enums are just ints, and any value is allowed, and in thrift python when deserializing such an enum with an invalid value, we explicitly represent it as `BadEnum`, but when converting a thrift py struct to thrift python, instead it just throws an error.

This is bad behaviour since it can lead to exceptions thrown for users when they don't expect them to be thrown. This updates the converter to instead wrap these wrong enums with `BadEnum`, and changes thrift python enums to accept `BadEnum` as an explicitly allowed input.

Reviewed By: junzh0u

Differential Revision: D39372476

fbshipit-source-id: bba0f58d32ee084fef3daad3525da945661cb8e5