Implement interaction factory functions in thrift python
Summary:
Adds support for interaction factory functions in thrift python, this is done by creating an interaction, then passing it to the omni client along with a normal request to set the newly created interaction as a factory.

Note that if client side RPCOptions were implemented then this would be able to be completed in a much better manner since it would involve just setting the interaction number on the RPCOptions passed in from the client, so once these are implemented I'll go back and improve this implementation.

Reviewed By: junzh0u

Differential Revision: D38848430

fbshipit-source-id: 1d914defba68f00d3c60930373fb115bdeabb0ac