semdiff: remove autocargo ignore directive
Summary:
D39412960 gets us most of the way towards being able to successfully execute `cargo check` on the hack workspace.

two problems remain:

(1) a link error in `hack_macro` for `-lsqlite3`
  - this issue was resolved by executing `sudo dnf install sqlite-devel`

(2) an error in `hackc` whereby we use an undeclared crate `sem_diff`. removing `autocargo = {"ignore_rule": True}` from the `sem_diff/TARGETS` file resolves that error and that's what this diff does.

i don't know if this is the right fix so consider this offered :)

Reviewed By: aorenste

Differential Revision: D39413895

fbshipit-source-id: 6948986f4084c8437c05ed4e61c2563db46c907a