Add enforceability calculation to decls for implicit pessimisation
Summary:
Pessimisation refers to weakening some types to be like-types (~t). If we pessimise enough types, then all classes and functions can be made to support the dynamic type. This in turn will let us freely use dynamic and like types to remove FIXMEs and UNSAFE CASTS.

We have been experimenting with explicitly pessimising code and seeing if we can type check it. This diff adds implicit pessimisation to the decl service. This means that when the type checker asks for a decl, it gets a pessimised version of the type of the decl, even though the source code itself has no like types. The implicit pessimisation strategy here is that types enforced by hhvm are unchanged (since they can't be wrong at runtime w/o encountering an exception), but unenforced property and method/function returns are pessimised (since hhvm will not catch it if they are wrong).

The Decl_enforceable.ml module is mostly a copy of the Typing_enforceability.ml module. However, they consume different environments, and the Typing_enforceability.ml module adds fanout edges, which the decl provider must not do.

Reviewed By: CatherineGasnier

Differential Revision: D39314187

fbshipit-source-id: 46611630dd86140d9656fadea6c195368384c047