unmarhsal refactor: remove Reverse_64()
Summary: `Reverse_64` reverses the order of 8 bytes of a slice in place. it can thus be eliminated in favor of the more general `reverse` function of the intrinsic type `slice`. we can go further though and write simpler, equivalent code in terms of `f64::from_le_bytes` and `f64::from_be_bytes`  which is what we do here.

Reviewed By: edwinsmith

Differential Revision: D39387339

fbshipit-source-id: 66b228cd7e6d088df12960fb7571761cec5344b0