Disable log bridge while running unit tests.
Summary:
Follow up on D39393506 (https://github.com/facebook/hhvm/commit/b6a998ea14bbb538f082ef6c55fdcd4fe27159d7) by actually using the option to disable the log bridge while running unit tests (for both the test runner and the test instances themselves, as well as any servers that the test run starts).

I originally thought that we'd want this option to be registered by fb-specific code (since the log bridge isn't shipped to open source), but having written this diff I think we should leave it where it is: this way we can pass the option in both internal and oss builds without triggering warnings. It's OK that the option is a noop in open source builds.

Reviewed By: oulgen

Differential Revision: D39454080

fbshipit-source-id: e788cdb6a58a092a0e13dda62288abc9f3fa031a