Mark all functions safe
Summary: Push all unsafety down to statement level and add comments explaining why these operations should be safe. Remove serialization of blocks with INFIX_TAG (which would require an `unsafe` block), since they only appear in closures and therefore shouldn't be reachable. Use `MaybeUninit` in `entries` to make explicit the invariant we rely on (we can only assume an element is initialized if its corresponding bit is set in `present`).

Reviewed By: shayne-fletcher

Differential Revision: D39443762

fbshipit-source-id: 0db2b4c2532865efd5267ddfcfbcb96749f07376