use containers for aliased_namespaces and include_roots
Summary:
Currently, the list of aliased_namespaces and include_roots are passed
from C++ to Rust as JSON-encoded strings. The Rust code is overly general,
and allows any named setting to be set this way, and it has workarounds for
the Hack JSON encoder which is no longer used. (C++ uses folly::dynamic).

Instead, pass these settings as a vector of string pairs, which is easy to encode in
the cxx bridge. This also reduces HPHP::compilers_start() to a NOP in HHVM,
so delete it.

Reviewed By: aorenste

Differential Revision: D39313582

fbshipit-source-id: 3b387c218bd8992e1e86c4db4fa220a3e5ff625d