modify fanout API to support edges-mode
Summary: This diff introduces an FFI for `hh_fanout` that the typechecker workers call into via `typing_deps.ml`. This is just the API; right now the methods are all stubbed and I've scattered quite a few TODOs that we'll need to address before edge files are actually being written. This API will be used for typechecker workers to create or update the depgraph.

Reviewed By: CatherineGasnier

Differential Revision: D36885005

fbshipit-source-id: 3be310d6b384454e746bfe96759582da9e3808ca