Make impls of EqModuloPos{,AndReason} less incorrect for {Hash,Index}{Map,Set}
Summary: Previously, we used `.iter().zip()`, which is clearly wrong: we should consider HashMaps and IndexMaps equal even when their iteration order differs. Instead, iterate over the keys, look up that key in the other map, and compare the value using `eq_modulo_pos{,_and_reason}`. This will not correctly handle maps containing keys which differ only in their positions or reasons, but we do not construct maps with keys containing positions or reasons, so this should be okay for now.

Reviewed By: ljw1004

Differential Revision: D39485237

fbshipit-source-id: 7f835c0fcbfe60cd91ea5d1b2ca4563438ad8d6b