HHVM admin server: specify charset in responses
Summary:
When the charset is not specified in the Content-Type header it is ambiguous how the bytes should be interpreted.

In the absence of a specific indication, some HTTP client stacks fall back to heuristics to try to parse the response, which can lead to performance issues.

This diff changes the Content-Type header for the admin server from `Content-Type: text/plain` to `Content-Type: text/plain; charset=utf-8` to explicitly specify how the response should be interpreted.

Reviewed By: binliu19

Differential Revision: D39484470

fbshipit-source-id: d2ba318bcfbd74473e60c1a8132efe41869d443e