Refining identifier entity/constraint
Summary:
In view of us wanting to add support for inheritance, we refine `identifier_entity`. In essence, we now model class constants `\C::DICT` as `(Some "\C", "DICT")` and global constants "DICT" as `(None, "DICT")`. This will allow us to access the constraints for  `\C` to resolve inheritance.

We now also distinguish the entity that we use to model for `D extends C` statements from the entity that we used for class/global constants.

Reviewed By: madgen

Differential Revision: D39432664

fbshipit-source-id: 9ac2523bcd0a870054da37880c98392d269e5135