Fixing `non_found` bug via inheritance
Summary:
Currently an analysis of the file

  <?hh

  class C {
    const dict<string, mixed> DICT = dict['a' => 42];
  }

  class D extends C {
  }

  function main(): void {
    idx(D::DICT, 'b');
  }

yields the error

  Uncaught exception:

    Not_found

The exception is caused by the call `SMap.find const_ident current_constraint_map` in `hips_solver.ml`. The issue is that the use of the identifier `D::DICT` in `main` lets the analysis assume that our constraint map has an entry at `D::DICT`. Since in `D` we inherit `DICT` from `C`, instead of defining it, this is a false assumption.

In this diff we fix above bug by using the `ClassExtends` constraint to look for constraints at `C::DICT`, instead of  `D::DICT`.

Merging copied code in a follow-up diff.

Reviewed By: madgen

Differential Revision: D39432987

fbshipit-source-id: ba671e7fcfd32ab45dbe5d342a6c48d0f595b9df