Fix resumetc and leavetc on ARM
Summary:
This diff changes the implementation of resumetc and leavetc on ARM.
It makes resumetc jump to the TC, instead of calling it.  The return
address from the call was being lost on ARM, as the LR wasn't saved
anywhere.  This diff also changes leavetc jump to the enterTCExit
unique stub, instead of doing a ret that was supposed to always go to
that stub (but wasn't, because LR wasn't saved).

Reviewed By: ricklavoie

Differential Revision: D39493113

fbshipit-source-id: 0cfe675a0aa608f923aba172922337a64f966b51