Include default values in runtime struct schema
Summary:
Filling in https://github.com/facebook/fbthrift/blob/main/thrift/lib/thrift/schema.thrift

Adds mechanism for interning values. These values are scoped to the program. Unlike the rest of these consts, these values require codegen in each target language for the resolution function. This diff adds such a function in C++, scoped as part of the program _constants struct to help the user get to the right scope.

Because the intern list is stored using AnyValue, which depends on some Thrift types, there has to be a bootstrapping mechanism to enable compiling those types without the interning dependencies. An opt-in approach is chosen for OSS-compatibility.

Reviewed By: Alfus

Differential Revision: D39367187

fbshipit-source-id: 093548748ec9395191efabd8402db74ffe27af4a