Handle legacy dynamic
Summary: In explicit pessimisation, `dynamic` parameters get replaced with `supportdyn<mixed>` so that we can call the function without having to upcast SDT arguments. Inside the function, `$param_name = $param_name upcast dynamic` gets added so that we can still perform dynamic operations on the parameter. For implicit pessimisation, we update the decl_provider to do the replacement of the parameter, which allow the function to be called. Instead of also adding the upcasts, we simply don't change the `dynamic` hint when checking the body of the function.

Reviewed By: vassilmladenov

Differential Revision: D39465977

fbshipit-source-id: 4d317715365be260b4ac0d07731e30711b390e1d