Revert back to a spinlock in DigestBuilder
Summary:
Counters are used in several latency-critical codepaths, for which a preemption (either through `yield` or a futex) can be very damaging.

D31654923 switched the spinlock to a `SharedMutex`, which is very eager to yield the CPU on contention, but that comes at the cost of tail latency. Instead, revert to a spinlock, but try to remove sources of contention by making the critical section very unlikely to be preempted.

Note that the mutex is per-cpu, so contention can happen only if we have a stale `cachedCurrent()`, or the thread migrates while in the critical section, or `build()` is running (but the critical section of that is also very small).

Reviewed By: LoganEvans

Differential Revision: D39521284

fbshipit-source-id: 474f5b486d5b2f4824de2beb27ad5a53db797573