make resource pool notification exception-safe
Summary:
The correct accounting of ParallelConcurrencyController depends on the ```onRequestFinished()``` callback to be called correctly in request destruction.

This callback now is only invoked in HandlerCallback destruction, which means that if an exception happens before that, the callback cannot be made correctly.

This diff set a contract that when HandlerCallback is constructed, it will "move" the callback pointers from the ServerRequest, which means that if the pointers are not moved, something bad must have happened before the construction. In that case we should also invoke the finish callback in the destructor of ServerRequest

Reviewed By: roddym

Differential Revision: D39518284

fbshipit-source-id: 2c00f640fa1c464b42b2fdfe34eeb88e2041959a