Build and test hhvm_clang in GitHub Actions (#9200)
Summary:
This PR adds the package `hhvm_clang` to the build matrix of GitHub Actions so that GitHub Action will try to build and test HHVM on Clang. It is currently marked as continue-on-error, because building `hhvm_clang` is expected to fail given that we have not fully ported HHVM OSS to Clang yet.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9200

Test Plan: GitHub Actions should build and test both `hhvm_clang` and `hhvm` packages, even though failures in `hhvm_clang` would not fail the entire GitHub Action workflow.

Reviewed By: alexeyt

Differential Revision: D39530989

Pulled By: Atry

fbshipit-source-id: e4ae82db532381ccd4ab43a1ee495f0f48ad7141