resolve `require class` eagerly when performing typing_extends checks for traits
Summary:
To avoid cycles in dependency folding, elements imported in a trait via a `require class` constraint are not flattened in the decls of the trait, and must explicitly searched for.

With this diff, if a trait T:
- implements interfaces that define the element `foo`
- does not define `foo` itself
- but `require class C`, and `C` defines `foo`

then we eagerly fetch the element `foo` from the required class when performing the `typing extends` checks.

Reviewed By: vsiles

Differential Revision: D39543345

fbshipit-source-id: c304a16cd484cc0ddc304a41a6bb9452e69ebe43