Remove two hot loads in readXXXBegin. (#527)
Summary:
X-link: https://github.com/facebook/fbthrift/pull/527

Remove one or two hot loads in readXXXBegin. In the 'likely' (else) side of the IF the two
return values are unused. By duplicating and sinking the code we allow the
compiler to inline into the hot path and eliminate the two hot loads.

Reviewed By: LoganEvans

Differential Revision: D39599564

fbshipit-source-id: 53d532a977d70f487923729c140f9e57272667b0