Restrict EqModuloPos impl for hash collections to collections with string keys
Summary:
The impl we have for EqModuloPos does not compare the keys of hash maps or the elements of hash sets in a position-insensitive fashion, so it won't work correctly for types like `IndexSet<UserAttribute<BPos>>`.

I'm not sure how to write EqModuloPos in an efficient way that compares keys without considering position, so I'm suggesting that we restrict these impls to collections with key types implementing AsRef<str>. This will include String, Symbol, MethodName, PropName, etc, but should exclude any type that contains a position or a reason. Every type which currently implements EqModuloPos and contains a {Hash,Index}{Set,Map} uses a key type that implements AsRef<str>.

Reviewed By: ljw1004

Differential Revision: D39488687

fbshipit-source-id: 5a9b31c9204aae70c946847e52ffadb396a1a75f