encode_kind(), encode_entry() helpers
Summary:
hint_to_type_constant() walks an ast::Hint and directly encodes it as "type constant",
which is nothing more than a weakly typed `dict` constant with stringly typed fields
and sub-dicts. In hhvm this is known as a TypeStructure. TypeConstants/Structures
are used to express the RHS of type aliases, class type constants, the RHS of is/as
operators, and reified generic types.

Now that we have a well defined enum for TypeStructureKind we can start using it
during emission:
* emit_kind() takes a TypeStructureKind and emits a `"kind" => int` entry for it.
callers can pass in an enum instead of an ad-hoc string that we convert to the enum.
* emit_entry() is a new utility used to emit all the stringly-typed fields.

There is an underlying schema to TypeStructure dicts, but AFAIK it is not expressed
in any strongly typed ADT nor is it documented. We leave that for future work.

Reviewed By: aorenste

Differential Revision: D39617042

fbshipit-source-id: b9a794d1b3b6ee70614a043f112baa7b7543f02c