Try harder to detect facebook builds in run.php
Summary: We currently add the 'facebook' category of tests to 'all' if hphp/facebook/test exists. When we run hhvm.no_facebook=1 builds locally we can't successfully run those tests, but the hphp/facebook/test directory has not been stripped. Refactor the check to first look for the hphp/facebook/test directory, but if it is found also use the runif machinery to check if the 'facebook' extension is available.

Reviewed By: ricklavoie

Differential Revision: D39645433

fbshipit-source-id: 2bf89ff312c47b6edb19e350d8d339db7fe97655