Fix ext_hash unit test in hhvm.no_facebook=1 builds
Summary:
This test was making different assertions about what hashing with adler32 and crc32b should produce in FB and non-FB builds, but they actually produce the same values and I see noevidence that they ever produced different values. Stop having different assertions, and just dump the hash value rather than comparing it to an expectation encoded in the test and dumping a bool (so that if it fails we can see what value was actually produced instead of a 'true' changing to a 'false').

In the test_furchash portion, fix the FB code path to no longer use HH\Lib\Legacy_FIXME\eq and the non-FB code path that we forgot to codemod from == to HH\Lib\Legacy_FIXME\eq. They should both use === and compare to an integer (the type furchash_hphp_ext actually returns) not an intish string.

Echo headers at the start of each test helper function so the test output/expectations are easier to navigate.

Finally, make some fixes to formattign and typos.

This test was previously failing in hhvm.no_facebook=1 builds, now it passes in both kinds of builds.

Reviewed By: oulgen

Differential Revision: D39591758

fbshipit-source-id: 75e089e162d57971000dc776201ad7e4982d55e3