Stop using fb-only class in const object with native data test
Summary:
We need to exercise the separate code paths for const properties on a class with __NativeData, but in order to test all the behavior we want we need to be able to new instances of the class ourselvs and var_dump them without having to call a native constructor first. When I originally wrote this test in D16107454 (https://github.com/facebook/hhvm/commit/c8244ca9e023e3aa41c3221e95874f0cd2437da1) the only class I could find that satisfied those requirements was StringBuffer, but unfortunately that's defined in an FB-only extension and we want this test to run in all environments.

Fix this by making a dummy test class in the intrinsics extension that stores a single int value as __NativeData, allows us to extend it and has no constructor. Use this class in the test instead of StringBuffer.

Also clean up cruft from when I was originally writing this test (before using StringBuffer) thinking I could use DateTimeZone as the base class in the test.

Reviewed By: ricklavoie

Differential Revision: D39591760

fbshipit-source-id: 1e750279fa58974d4576f74a6b1884232cec598d