Enforce static property accesses
Summary:
This diff implements static property enforcement for modules V1.

- Refactors RaiseModulePropertyViolation to take a specialized extra-data containing all static information about properties, and adds a "is_static" boolean for erroring on static properties

We update the interpretor and jit to enforce property violations for static property accesses. These are split into enforcing the following bytecodes:

- CGetS: gets a static prop as a cell
- SetS: sets a static prop in a cell
- BaseSC: In member instructions, sets a static property as a base
- IssetS: check if a static property is set

We disable a few hhbbc optimizations related to nothrows when a static property is internal, since it may now throw. We can later improve this so that we still do the optimization when the static property's module matches the caller's module, and improve it if they do not match if were are 100% sure we will throw. For now though, considering the limited use of internal static properties (as of now, literally zero uses in the codebase), improving it to be even smarter may not be immediately valuable.

Reviewed By: oulgen

Differential Revision: D39362576

fbshipit-source-id: 0ecf24fcf6862266be40ec46c4d6efc592fe5b4d