pos_span_tiny fix
Summary: OCaml integers are 63-bit & signed. It does not offer a builtin unsigned  type. In `pos_span_tiny::PosSpanTiny` however we represent source locations as 63-bit *unsigned* integers. In these cases we often produce 63-bit integers that are > 2^62 - 1. If such an integer is encoded as an OCaml value & then naively decoded as if it were a signed 63-bit int, we produce the wrong result (the decoded value will be negative). This diff updates `pos_span_tiny::PosSpanTiny::from_ocaml_rep` to account for this.

Differential Revision: D39660472

fbshipit-source-id: 83bd6718e2c0970720a5bbc0d0ee75c226604861