Add multifile test to demonstrate error in decl comparison test
Summary:
There is a bug in our "rupro vs ocaml single file decl comparison binary". Namely, when working with "multifile single files", that is, a single file that takes advantage of our special multifile syntax to pretend to be multiple files, we get file order dependent behavior.

This diff merely adds a test that demonstrates this problem.

```
//// b1.php
<?hh

class B1 extends A1 {}

//// a1.php
<?hh

class A1 {
  public function foo(): void {}
}
```

In OCaml, when I parse and declare file b1.php, we query the provider for class A1, but because we haven't seen that file before, it doesn't exist in the provider and so we don't properly fold the decl for B1 with all of the information from A1.

The difference can be seen here:

```
File root|multifile_decl_ordering.php--b1.php
 --- ocaml
+++ rupro
@@ -1,23 +1,40 @@
 FoldedClass {
     name: "\\B1",
     pos: Pos(root|multifile_decl_ordering.php--b1.php, 3:7-9),
     kind: Cclass(
         Concrete,
     ),
+    substs: {
+        "\\A1": SubstContext {
+            subst: Subst(
+                {},
+            ),
+            class_context: "\\B1",
+            from_req_extends: false,
+        },
+    },
     ancestors: {
         "\\A1": A1,
     },
+    methods: {
+        "foo": FoldedElement {
+            flags: (empty),
+            origin: "\\A1",
+            visibility: Public,
+            deprecated: None,
+        },
+    },
     consts: {
         "class": ClassConst {
             is_synthesized: true,
             kind: CCConcrete,
             pos: Pos(root|multifile_decl_ordering.php--b1.php, 3:7-9),
             ty: HH\classname<this>,
             origin: "\\B1",
             refs: [],
         },
     },
     extends: {
         "\\A1",
     },
 }
```

However, if I reorder the files:

```
//// a1.php
<?hh

class A1 {
  public function foo(): void {}
}

//// b1.php
<?hh

class B1 extends A1 {}
```

The OCaml logic, when looking for A1 when parsing and declaring b1.php, will find A1's decl because it had previously parsed it and we will get the correct answer.

This test should be committed to prevent future regressions - the next diff fixes this behavior and deletes the corresponding exp file. This diff demonstrates the difference.

Reviewed By: panopticoncentral

Differential Revision: D39672217

fbshipit-source-id: c6d3c40e895b1f6bc50277116d73c0692cb3fb35