Copy intern crate into hack from relay
Summary:
This stack attempts to allow HackC to take a naming table to power decl directed bytecode.

However, in order to do so, this stack takes on a dependency on hackrs which in turn takes on the `intern` crate dependency from the `relay` directory. Despite being OSS, the relay directory and its crates are not published on crates.io which leads to our hhvm/hack OSS build breaking.

As a workaround, it was suggested to copy the `intern` crate into hack's own directory, removing the dependency on relay. When the `intern` crate from relay is published in crates.io and we can depend on that externally, we can delete this copy and hopefully prevent skew in the meantime.

This diff in the stack simply makes a copy of the crate to hack's util directory.

Reviewed By: shayne-fletcher

Differential Revision: D39397150

fbshipit-source-id: 31b1a60d7f85176f1e7235eae9e9a356ce792e4d