Change hack TARGETS' to point to copied intern crate
Summary:
This stack attempts to allow HackC to take a naming table to power decl directed bytecode.

However, in order to do so, this stack takes on a dependency on hackrs which in turn takes on the intern crate dependency from the relay directory. Despite being OSS, the relay directory and its crates are not published on crates.io which leads to our hhvm/hack OSS build breaking.

As a workaround, it was suggested to copy the intern crate into hack's own directory, removing the dependency on relay. When the intern crate from relay is published in crates.io and we can depend on that externally, we can delete this copy and hopefully prevent skew in the meantime.

This diff in the stack points all of the TARGETS under /hphp/hack/src/ to point to the `intern` crate in hack rather than relay's copy.

Reviewed By: shayne-fletcher

Differential Revision: D39397151

fbshipit-source-id: 1fefdcadd6cb1a9392b7b647588036e53f87adac