Make enum values available as const for typedef-ed enums
Summary:
Example problem:
In "original_def.thrift":
```
enum Letter {
 A = 0,
 B = 1,
 ...
}
```
In "alias.thrift":
```
include "original_def.thrift"
typedef original_def.Letter Letter
```
In "user.thrift":
```
include "alias.thrift"
struct User {
    1: alias.Letter favorite_letter = alias.Letter.A,
}
```
would yield warning (which, based on a comment right above the warning line, should be turned into an error in the future):
```
The identifier 'alias.Letter.A' is not defined yet.
Constants and enums should be defined before using them as default values.
```

Cause:
When handling a enum declaration, thrift parser adds all its enum values as constants to the scope cache, so that later when the identifiers are used as default values, the parser recognizes them and knows the corresponding values. But this doesn't happen when the enum is typedef-ed - it only happens for the original declaration of the enum.

Solution:
When handling type declarations, checks if it's a typedef of an enum, and if yes, do the same thing as when handling the original declaration (add all values as constants to scope cache).

Result:
Warning is gone.

Alternative:
Try searching for values of typedef-ed enums in `parsing_driver::find_const` lazily. But that might be more complex.

Reviewed By: vitaut

Differential Revision: D39641834

fbshipit-source-id: 8b54dad62bd68782649c5e05e6753fdaeb928f61