inline simple consts
Summary:
Textual allows fairly complex expressions - so instead of writing simple constants as:
```
n0 = hack_null()
ret n0
```
combine them to produce:
```
ret hack_null()
```
This is entirely for readability of the resulting Textual file.

We don't currently emit any constructs that would be considered "complex" like vec or dict but the code is there to keep emitting them as a two-step instruction.

Reviewed By: davidpichardie

Differential Revision: D39642270

fbshipit-source-id: dc6fcde2f71049009c7bf95a8c6a7b781d504076