Remove libgccgit (#9211)
Summary:
Previously I added `libgccgit` to the HHVM nix package, in order to add `/nix/store/ykgwb51j5235xdbkqnznm2haygknmwc6-libgccjit-11.3.0/lib/libatomic.so.1` to the library search path so that `find_library` could find `libatomic.so.1`'s full path. However, it is unnecessary because gcc also provides another `libatomic.so.1`, which is always available by using the `-latomic` flag, even when `libatomic.so.1` is not in the library search path.

This PR removes `libgccgit` and instead lets `gcc` directly link HHVM with gcc's built-in libatomic if `find_library` cannot find the full path of `libatomic.so.1`. This PR could avoid potential ABI incompatibility between gcc's and libgccgit's `libatomic.so.1` .

Pull Request resolved: https://github.com/facebook/hhvm/pull/9211

Test Plan: Both internal CI and GitHub Actions for gcc should pass. The GitHub Action for clang is expected to fail because we have not fully port HHVM OSS to clang.

Reviewed By: Wilfred

Differential Revision: D39643712

Pulled By: Atry

fbshipit-source-id: e455846f52d5afa447ff44b6f5dc3d877d4b007b