Make sync_* methods preceed unprefixed sync methods.
Summary:
Given a service method named `foo`, the current behavior is that the generated `foo` method in the `ServiceHandler` class calls `sync_foo`. This diff reverses the ordering so that `sync_foo` calls `foo` instead. This is safe because no one implements `sync_foo` yet.

This allows codemods that update direct calls of `foo` to `sync_foo`.

Reviewed By: iahs

Differential Revision: D39716124

fbshipit-source-id: d9462a1bf91db706f02dbbf7732ae2ce68753864