revise folly::is_constant_evaluated
Summary:
Since there is no single, uniform, correct way to backport it.

The new interface is:
```lang=c++
namespace folly {
constexpr bool is_constant_evaluated_or(/* constinit */ bool /* default-value */) noexcept;
}
```

We rename to `is_constexpr_evaluated_or` and we add mandatory fallback argument which will be returned if the code is not able to determine whether or not it is in a constant context.

Reviewed By: Gownta, DenisYaroshevskiy

Differential Revision: D39585698

fbshipit-source-id: 5f255940d5c1ee1a31dcaaf6bb7b7aafcecf64b4