Add StringData pretty printer, skeleton for ArrayData; move walkstk class
Summary:
This adds the pretty printer for the `HPHP::StringData` class. It also adds the `ArrayData` printer for the `kVecKind` (just the metadata until I implement the iterator), and moves the `walkstk` class to its own module and adds a `Command` base class to handle some of the boilerplate of adding the commands to the debugger.

The main purpose is to get a good framework going for adding additional printers and commands, and to be able to (re)load them in lldb for debugging/altering the scripts themselves as we develop them further.

Reviewed By: arnabde03

Differential Revision: D39641653

fbshipit-source-id: dff2e44b03f6dbbcec520cf99c925e9cd07ec071